refactor: extract storage blocks to models/storage_blocks.py
Move DownloadToS3Block, UploadToS3Block, and FileUploadBlock out of
block.py into a new models/storage_blocks.py, with block.py re-exporting
all three via a late import (matching the google_sheets_blocks and
pdf_fill_block pattern). Prune the block.py preamble imports that became
unused after the cut.
